Delegate calendar month generation to MonthConfig class.
