+++ library/src/test/java/com/kizitonwose/calenderview/CalenderViewTests.kt
@@ -1,11 +1,7 @@
 package com.kizitonwose.calenderview
 
 import androidx.recyclerview.widget.RecyclerView
-import com.kizitonwose.calendarview.model.CalendarMonth
-import com.kizitonwose.calendarview.model.DayOwner
-import com.kizitonwose.calendarview.model.OutDateStyle
-import com.kizitonwose.calendarview.model.ScrollMode
-import com.kizitonwose.calendarview.ui.MonthConfig
+import com.kizitonwose.calendarview.model.*
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.threeten.bp.DayOfWeek
@@ -24,10 +20,20 @@ class CalenderViewTests {
     private val firstDayOfWeek = DayOfWeek.MONDAY
 
     private val endOfRowConfig =
-        MonthConfig(OutDateStyle.END_OF_ROW, ScrollMode.PAGED, RecyclerView.HORIZONTAL, null)
+        MonthConfig(
+            OutDateStyle.END_OF_ROW,
+            ScrollMode.PAGED,
+            RecyclerView.HORIZONTAL,
+            null
+        )
 
     private val endOfGridConfig =
-        MonthConfig(OutDateStyle.END_OF_GRID, ScrollMode.PAGED, RecyclerView.HORIZONTAL, null)
+        MonthConfig(
+            OutDateStyle.END_OF_GRID,
+            ScrollMode.PAGED,
+            RecyclerView.HORIZONTAL,
+            null
+        )
 
 
     @Test
